requirements.txt
altgraph==0.17.4
et-xmlfile==1.1.0
macholib==1.16.3
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
